% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGV.R
\name{conditionLatentObs}
\alias{conditionLatentObs}
\title{Assign conditioning sets for the SGV approximation}
\usage{
conditionLatentObs(nID, coords_ord, N)
}
\arguments{
\item{nID}{N x k matrix of neighbor indices.}

\item{coords_ord}{N x 2 matrix of locations.}

\item{N}{Scalar; number of locations (observed only!).}
}
\value{
A matrix indicating whether the conditioning set for each location 
is on the latent process (y, \code{1}) or the observed values (z, \code{0}).
}
\description{
\code{conditionLatentObs} assigns q_y(i) vs q_z(i) following Section 5.1 
in Katzfuss and Guinness (2018). This function only needs to be run once 
per SGV analysis.
}
