% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v03_NNGP_neighbors.R
\name{displayNeighbors}
\alias{displayNeighbors}
\title{Plot neighbor sets for a Nearest-Neighbor Gaussian Process (NNGP)}
\usage{
displayNeighbors(ind, coords.sorted, neighbor_idx, proj)
}
\arguments{
\item{ind}{Integer index of the focal location (in the sorted order).}

\item{coords.sorted}{Numeric matrix (\eqn{M \times 2}) of spatial coordinates sorted
according to the ordering used in the NNGP construction. Output from \code{computeNeighbors}.}

\item{neighbor_idx}{Integer matrix (\eqn{M \times k}) where each row contains the indices
of the neighbors for a given location. Output from \code{computeNeighbors}.}

\item{proj}{Coordinate reference system (CRS) used for plotting. Can be an EPSG code
or a CRS object compatible with \code{sf}.}
}
\value{
A \code{ggplot2} object showing:
\itemize{
  \item All spatial locations (black markers)
  \item The focal location (red point)
  \item Its neighbors (blue points)
  \item A vertical reference line at the focal location's x-coordinate
}
}
\description{
Visualizes the neighbor set for a given location based on an ordered
Nearest-Neighbor Gaussian Process (NNGP) constructed using Vecchia's approximation.
}
\details{
The function assumes that coordinates and neighbor indices are based on an
ordered representation of the data (e.g., sorted by x-coordinate or another
ordering used in Vecchia's approximation).

The neighbor set for location \eqn{i} consists of previously ordered locations,
as defined by the NNGP construction.
}
\examples{
coords <- matrix(runif(40), ncol = 2)
res <- computeNeighbors(coords, k = 5, proj = 4326)
p <- displayNeighbors(
  ind = 10,
  coords.sorted = res$coords_sorted,
  neighbor_idx = res$neighbor_idx,
  proj = 4326
)
print(p)

}
\seealso{
computeNeighbors
}
\author{
Fabian Ketwaroo
}
