% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{inverseEigen}
\alias{inverseEigen}
\title{Calculate covariance elements based on eigendecomposition components}
\usage{
inverseEigen(eigen_comp1, eigen_comp2, eigen_comp3, which_Sigma)
}
\arguments{
\item{eigen_comp1}{N-vector; contains values of the log of the second
anisotropy eigenvalue for a set of locations.}

\item{eigen_comp2}{N-vector; contains values of the first eigenvector
component for a set of locations.}

\item{eigen_comp3}{N-vector; contains values of the second eigenvector 
component for a set of locations.}

\item{which_Sigma}{Scalar; one of \code{(1,2,3)}, corresponding to which
covariance component should be calculated (Sigma11, Sigma22, or Sigma12,
respectively).}
}
\value{
A correlation matrix for a fixed set of stations and fixed
parameter values.
}
\description{
\code{inverseEigen} calculates the inverse eigendecomposition -- in other
words, the covariance elements based on the eigenvalues and vectors (see
Paciorek and Schervish, 2006, for details on the parameterization). The 
function is coded as a \code{nimbleFunction} (see the \code{nimble} package) 
but can also be used as a regular R function.
}
\examples{
# TODO

}
