% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGV.R
\name{nimble_sparse_chol}
\alias{nimble_sparse_chol}
\title{nimble_sparse_chol}
\usage{
nimble_sparse_chol(i, j, x, n)
}
\arguments{
\item{i}{TODO}

\item{j}{TODO}

\item{x}{TODO}

\item{n}{TODO}
}
\description{
nimble_sparse_chol
}
