% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp2Scale.R
\name{nimble_sparse_cholesky}
\alias{nimble_sparse_cholesky}
\title{nimble_sparse_chol}
\usage{
nimble_sparse_cholesky(i, j, x)
}
\arguments{
\item{i}{Vector of row indices.}

\item{j}{Vector of column indices.}

\item{x}{Vector of values in the matrix.}
}
\description{
nimble_sparse_chol
}
