% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{nsgpModel}
\alias{nsgpModel}
\title{NIMBLE code for a generic nonstationary GP model}
\usage{
nsgpModel(tau_model = "constant", sigma_model = "constant",
  Sigma_model = "constant", mu_model = "constant",
  likelihood = "fullGP", coords, z, constants = list(),
  returnModelComponents = FALSE, ...)
}
\arguments{
\item{tau_model}{Character; specifies the model to be used for the log(tau) 
process. Options are \code{"logLinReg"} (log-linear regression), 
\code{"mixComp"} (mixture component representation), \code{"GP"} (stationary
Gaussian process), and \code{"approxGP"} (approximation to a Gaussian 
process).}

\item{sigma_model}{Character; specifies the model to be used for the 
log(sigma) process. See \code{tau_model} for options.}

\item{Sigma_model}{Character; specifies the model to be used for the 
Sigma anisotropy process. Options are \code{"covReg"} (covariance 
regression), \code{"compReg"} (componentwise regression), 
\code{"npMixComp"} (nonparameteric regression via the mixture component
approach), \code{"npGP"} (nonparameteric regression via a stationary 
Gaussian process), or \code{"npApproxGP"} (nonparameteric regression via an
approximation to a stationary Gaussian process).}

\item{mu_model}{TODO}

\item{likelihood}{TODO}

\item{coords}{TODO}

\item{z}{TODO}

\item{constants}{TODO}

\item{returnModelComponents}{TODO}

\item{...}{TODO}
}
\value{
A \code{nimbleCode} object.
}
\description{
TODO: add documentation
}
\examples{
# TODO

}
