% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{nsgpPredict}
\alias{nsgpPredict}
\title{Posterior prediction for the NSGP}
\usage{
nsgpPredict(model, samples, coords.predict, predict.y = TRUE, constants,
  seed = 0, ...)
}
\arguments{
\item{model}{TODO}

\item{samples}{TODO}

\item{coords.predict}{TODO}

\item{predict.y}{TODO}

\item{constants}{TODO}

\item{seed}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
\code{nsgpPredict} conducts posterior prediction for MCMC samples generated
using nsgpModel.
}
\examples{
# TODO

}
