% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{orderCoordinatesMMD}
\alias{orderCoordinatesMMD}
\title{Order coordinates according to a maximum-minimum distance criterion.}
\usage{
orderCoordinatesMMD(s, exact = FALSE)
}
\arguments{
\item{s}{N x 2 array of N 2-dimensional (x,y) spatial coordinates.}

\item{exact}{Logical; \code{FALSE} uses a fast approximation to MMD ordering 
(and is almost always recommended), while \code{TRUE} uses exact MMD 
ordering but is infeasible for large number of locations.}
}
\value{
A list with two components: (1) an N x 2 array containing the 
same spatial coordinates, ordered by MMD, and (2) the same thing, but with 
any NA values removed.
}
\description{
\code{orderCoordinatesMMD} orders an array of (x,y) spatial coordinates 
according to the "maximum minimum distance" (MMD), as described in Guinness, 
2018. (Points are selected to maximize their minimum distance to already-
selected points).
}
\examples{
# TODO

}
