% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGV.R
\name{R_sparse_chol}
\alias{R_sparse_chol}
\title{R_sparse_chol}
\usage{
R_sparse_chol(i, j, x, n)
}
\arguments{
\item{i}{Vector of row indices.}

\item{j}{Vector of column indices.}

\item{x}{Vector of values in the matrix.}

\item{n}{Length of the vector}
}
\description{
R_sparse_chol
}
