% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGV.R
\name{R_sparse_crossprod}
\alias{R_sparse_crossprod}
\title{nimble_sparse_crossprod}
\usage{
R_sparse_crossprod(i, j, x, z, n, subset = -1, transp = 1)
}
\arguments{
\item{i}{TODO}

\item{j}{TODO}

\item{x}{TODO}

\item{z}{TODO}

\item{n}{TODO}

\item{subset}{TODO}

\item{transp}{TODO}
}
\description{
nimble_sparse_crossprod
}
