% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGV.R
\name{R_sparse_solve}
\alias{R_sparse_solve}
\title{nimble_sparse_solve}
\usage{
R_sparse_solve(i, j, x, z)
}
\arguments{
\item{i}{TODO}

\item{j}{TODO}

\item{x}{TODO}

\item{z}{TODO}
}
\description{
nimble_sparse_solve
}
