% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp2Scale.R
\name{R_sparse_solveMat}
\alias{R_sparse_solveMat}
\title{nimble_sparse_crossprod}
\usage{
R_sparse_solveMat(i, j, x, z, transp = 1)
}
\arguments{
\item{i}{Vector of row indices.}

\item{j}{Vector of column indices.}

\item{x}{Vector of values in the matrix.}

\item{z}{Vector to calculate the cross-product with.}

\item{transp}{Optional indicator of using the transpose}
}
\description{
nimble_sparse_crossprod
}
