% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGV.R
\name{R_sparse_tcrossprod}
\alias{R_sparse_tcrossprod}
\title{nimble_sparse_tcrossprod}
\usage{
R_sparse_tcrossprod(i, j, x, subset = -1)
}
\arguments{
\item{i}{TODO}

\item{j}{TODO}

\item{x}{TODO}

\item{subset}{TODO}
}
\description{
nimble_sparse_tcrossprod
}
