% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp2Scale.R
\name{rmnorm_gp2Scale}
\alias{rmnorm_gp2Scale}
\title{Function for the evaluating the SGV approximate density.}
\usage{
rmnorm_gp2Scale(n, mean, Cov, N, Nnz)
}
\arguments{
\item{mean}{Vector of mean valiues}

\item{N}{Number of measurements in x}

\item{x}{Vector of measurements}

\item{Csparse}{Matrix of size N x N; sparse kernel}

\item{Ccore}{Matrix of size N x N; core kernel}

\item{lognuggetSD}{Vector of log nugget (measurement error) standard deviation.}

\item{log}{Logical; should the density be evaluated on the log scale.}
}
\value{
Not applicable.
}
\description{
\code{dmnorm_gp2Scale} (and \code{rmnorm_gp2Scale}) calculate the usual Gaussian
likelihood for a fixed set of parameters (but with sparse matrices). Finally,
the distributions must be registered within \code{nimble}.
}
