% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNGP.R
\name{rmnorm_nngp}
\alias{rmnorm_nngp}
\title{Function for the evaluating the NNGP approximate density.}
\usage{
rmnorm_nngp(n, mean, AD, nID, N, k)
}
\arguments{
\item{n}{N-vector of data.}

\item{mean}{N-vector with current values of the mean}

\item{AD}{N x (k+1) matrix; the first k columns are the 'A' matrix, and the
last column is the 'D' vector.}

\item{nID}{N x k matrix of neighbor indices.}

\item{N}{Scalar; number of data measurements.}

\item{k}{Scalar; number of nearest neighbors.}
}
\value{
The NNGP approximate density.
}
\description{
\code{dmnorm_nngp} (and \code{rmnorm_nngp}) calculate the approximate NNGP
likelihood for a fixed set of parameters (i.e., A and D matrices). Finally,
the distributions must be registered within \code{nimble}.
}
