% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGV.R
\name{rmnorm_sgv}
\alias{rmnorm_sgv}
\title{Function for the evaluating the SGV approximate density.}
\usage{
rmnorm_sgv(n, mean, U, N, k)
}
\arguments{
\item{n}{TODO}

\item{mean}{TODO}

\item{U}{TODO}

\item{N}{TODO}

\item{k}{TODO}
}
\value{
TODO
}
\description{
\code{dmnorm_sgv} (and \code{rmnorm_sgv}) calculate the approximate SGV
likelihood for a fixed set of parameters (i.e., the U matrix). Finally,
the distributions must be registered within \code{nimble}.
}
