% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGV.R
\name{nimble_sparse_chol}
\alias{nimble_sparse_chol}
\title{nimble_sparse_chol}
\usage{
nimble_sparse_chol(i, j, x, n)
}
\arguments{
\item{i}{Vector of row indices.}

\item{j}{Vector of column indices.}

\item{x}{Vector of values in the matrix.}

\item{n}{Length of the vector}
}
\description{
nimble_sparse_chol
}
