% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGV.R
\name{nimble_sparse_solve}
\alias{nimble_sparse_solve}
\title{nimble_sparse_solve}
\usage{
nimble_sparse_solve(i, j, x, z)
}
\arguments{
\item{i}{Vector of row indices.}

\item{j}{Vector of column indices.}

\item{x}{Vector of values in the matrix.}

\item{z}{Vector to calculate the cross-product with.}
}
\description{
nimble_sparse_solve
}
