% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGV.R
\name{nimble_sparse_tcrossprod}
\alias{nimble_sparse_tcrossprod}
\title{nimble_sparse_tcrossprod}
\usage{
nimble_sparse_tcrossprod(i, j, x, subset)
}
\arguments{
\item{i}{Vector of row indices.}

\item{j}{Vector of column indices.}

\item{x}{Vector of values in the matrix.}

\item{subset}{Optional vector of rows to include in the calculation.}
}
\description{
nimble_sparse_tcrossprod
}
