% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGV.R
\name{R_sparse_crossprod}
\alias{R_sparse_crossprod}
\title{nimble_sparse_crossprod}
\usage{
R_sparse_crossprod(i, j, x, z, n, subset = -1, transp = 1)
}
\arguments{
\item{i}{Vector of row indices.}

\item{j}{Vector of column indices.}

\item{x}{Vector of values in the matrix.}

\item{z}{Vector to calculate the cross-product with.}

\item{n}{Length of the vector}

\item{subset}{Optional vector of rows to include in the calculation.}

\item{transp}{Optional indicator of using the transpose}
}
\description{
nimble_sparse_crossprod
}
