% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp2Scale.R
\name{rmnorm_gp2Scale}
\alias{rmnorm_gp2Scale}
\title{Function for the evaluating the SGV approximate density.}
\usage{
rmnorm_gp2Scale(n, mean, Cov, N, Nnz)
}
\arguments{
\item{n}{Number of realizations to generate}

\item{mean}{Vector of mean values}

\item{Cov}{Matrix of size N x N; sparse kernel}

\item{N}{Number of measurements in x}

\item{Nnz}{Number of measurements in x}
}
\value{
Not applicable.
}
\description{
\code{dmnorm_gp2Scale} (and \code{rmnorm_gp2Scale}) calculate the usual Gaussian
likelihood for a fixed set of parameters (but with sparse matrices). Finally,
the distributions must be registered within \code{nimble}.
}
